# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Дошка оголошень API",
      "version": "1.0.0"
    },
    "paths": {
      "/api/register": {
        "post": {
          "summary": "Реєстрація користувача",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "201": { "description": "Успішна реєстрація" },
            "400": { "description": "Користувач вже існує" }
          }
        }
      },
      "/api/login": {
        "post": {
          "summary": "Логін користувача",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Успішний логін" },
            "401": { "description": "Невірний логін або пароль" }
          }
        }
      },
      "/api/ads": {
        "get": {
          "summary": "Отримати всі оголошення",
          "security": [{ "bearerAuth": [] }],
          "responses": {
            "200": { "description": "Список оголошень" },
            "401": { "description": "Неавторизовано" }
          }
        },
        "post": {
          "summary": "Створити нове оголошення",
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": { "type": "string" },
                    "description": { "type": "string" },
                    "price": { "type": "number" }
                  },
                  "required": ["title", "description", "price"]
                }
              }
            }
          },
          "responses": {
            "201": { "description": "Оголошення створено" },
            "401": { "description": "Неавторизовано" }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }
  